working gui and joint notes

+++ src/DAL/RealNoteDAO.java
@@ -80,8 +80,33 @@ public class RealNoteDAO implements NoteDAO<NoteDTO> {
         return noteDTO;
     }
 
+//    @Override
+//    public void update(NoteDTO noteDTO, NoteDTO newNoteDTO) throws SQLException {
+//        String sql = "UPDATE notes SET content = ?, year = ?, month = ?, day = ?, isJoint = ?, isDone = ? WHERE id = ?";
+//        PreparedStatement update = connection.prepareStatement(sql);
+//        connection.setAutoCommit(false);
+//        update.setString(1, newNoteDTO.getContent());
+//        update.setString(2, String.valueOf(newNoteDTO.getYear()));
+//        update.setString(3, String.valueOf(newNoteDTO.getMonth()));
+//        update.setString(4, String.valueOf(newNoteDTO.getDay()));
+//        if (newNoteDTO.IsJoint()) {
+//            update.setString(5, "1");
+//        } else {
+//            update.setString(5, "0");
+//        }
+//        if (newNoteDTO.IsDone()) {
+//            update.setString(6, "1");
+//        } else {
+//            update.setString(6, "0");
+//        }
+//        update.setString(7, noteDTO.getId());
+//        update.executeUpdate();
+//        connection.commit();
+//        connection.setAutoCommit(true);
+//    }
+
     @Override
-    public void update(NoteDTO noteDTO, NoteDTO newNoteDTO) throws SQLException {
+    public void update(NoteDTO newNoteDTO) throws SQLException {
         String sql = "UPDATE notes SET content = ?, year = ?, month = ?, day = ?, isJoint = ?, isDone = ? WHERE id = ?";
         PreparedStatement update = connection.prepareStatement(sql);
         connection.setAutoCommit(false);
@@ -99,7 +124,7 @@ public class RealNoteDAO implements NoteDAO<NoteDTO> {
         } else {
             update.setString(6, "0");
         }
-        update.setString(7, noteDTO.getId());
+        update.setString(7, newNoteDTO.getId());
         update.executeUpdate();
         connection.commit();
         connection.setAutoCommit(true);
@@ -107,18 +132,29 @@ public class RealNoteDAO implements NoteDAO<NoteDTO> {
 
     @Override
     public void delete(NoteDTO noteDTO) throws SQLException {
-        String sql = "DELETE FROM notes WHERE id = ?";
-        PreparedStatement preparedStatement = connection.prepareStatement(sql);
-        connection.setAutoCommit(false);
-        preparedStatement.setString(1, noteDTO.getId());
-        preparedStatement.executeUpdate();
-        connection.commit();
-        sql = "DELETE FROM jointNotes WHERE noteId = ?";
-        PreparedStatement deepDelete = connection.prepareStatement(sql);
-        preparedStatement.setString(1, noteDTO.getId());
-        preparedStatement.executeUpdate();
-        connection.commit();
+        if (get(noteDTO.getId()).getUserId().equals(noteDTO.getUserId())) {
+            connection.setAutoCommit(false);
+            String sql = "DELETE FROM notes WHERE id = ?";
+            PreparedStatement preparedStatement = connection.prepareStatement(sql);
+            preparedStatement.setString(1, noteDTO.getId());
+            preparedStatement.executeUpdate();
+            connection.commit();
+            sql = "DELETE FROM jointNotes WHERE noteId = ?";
+            PreparedStatement deepDelete = connection.prepareStatement(sql);
+            deepDelete.setString(1, noteDTO.getId());
+            deepDelete.executeUpdate();
+            connection.commit();
+        } else {
+            connection.setAutoCommit(false);
+            String sql = "DELETE FROM jointNotes WHERE noteId = ? AND userId = ?";
+            PreparedStatement preparedStatement = connection.prepareStatement(sql);
+            preparedStatement.setString(1, noteDTO.getId());
+            preparedStatement.setString(2, noteDTO.getUserId());
+            preparedStatement.executeUpdate();
+            connection.commit();
+        }
         connection.setAutoCommit(true);
+
     }
